chore: rename `<name>{=> /index}.ts` and `<name>{=> /style}.css` already having `<name>/`
